requirements.txt
asgiref==3.3.1
Django==3.1.7
django-bower==5.2.0
django-chartjs==2.2.1
django-crispy-forms==1.11.2
django-jsonfield==1.4.1
django-nvd3==0.9.7
djangorestframework==3.12.4
djcacheutils==3.0.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-memcached==1.59
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
Unidecode==1.2.0
